Move parameters.h to src since it is only used internally
