+++ src/parameters.c
@@ -1,9 +1,10 @@
-#include <cgreen/parameters.h>
 #include <cgreen/vector.h>
 #include <stdlib.h>
 #include <stdbool.h>
 #include <string.h>
 #include <ctype.h>
+
+#include "parameters.h"
 
 #ifdef __cplusplus
 namespace cgreen {
